second review, add precode_test file
