+++ precode.go
@@ -2,13 +2,8 @@ package main
 
 import (
 	"net/http"
-	"net/http/httptest"
 	"strconv"
 	"strings"
-	"testing"
-
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
 )
 
 var cafeList = map[string][]string{
@@ -47,41 +42,4 @@ func mainHandle(w http.ResponseWriter, req *http.Request) {
 
 	w.WriteHeader(http.StatusOK)
 	w.Write([]byte(answer))
-}
-
-func TestMainHandlerWhenCountMoreThanTotal(t *testing.T) {
-	totalCount := 4
-	req := httptest.NewRequest("GET", "/cafe?count=7&city=moscow", nil)
-
-	responseRecorder := httptest.NewRecorder()
-	handler := http.HandlerFunc(mainHandle)
-	handler.ServeHTTP(responseRecorder, req)
-
-	body := responseRecorder.Body.String()
-	list := strings.Split(body, ",")
-
-	require.Equal(t, http.StatusOK, responseRecorder.Code)
-	assert.Len(t, list, totalCount)
-}
-
-func TestMainHandlerHttp200(t *testing.T) {
-	req := httptest.NewRequest("GET", "/cafe?count=2&city=moscow", nil)
-
-	responseRecorder := httptest.NewRecorder()
-	handler := http.HandlerFunc(mainHandle)
-	handler.ServeHTTP(responseRecorder, req)
-
-	require.Equal(t, responseRecorder.Code, http.StatusOK)
-	assert.NotEmpty(t, responseRecorder.Body.String())
-}
-
-func TestMainHandlerWrongCity(t *testing.T) {
-	req := httptest.NewRequest("GET", "/cafe?count=2&city=saratov", nil)
-
-	responseRecorder := httptest.NewRecorder()
-	handler := http.HandlerFunc(mainHandle)
-	handler.ServeHTTP(responseRecorder, req)
-
-	require.Equal(t, responseRecorder.Code, http.StatusBadRequest)
-	assert.Equal(t, responseRecorder.Body.String(), "wrong city value")
 }
